MOTOR-40 Start pulling framework-agnostic aspects of Motor into core.py.
Preliminary to supporting Tornado and asyncio frameworks.
