+++ motor/motor_common.py
@@ -1,4 +1,4 @@
-# Copyright 2014 MongoDB, Inc.
+# Copyright 2011-2014 MongoDB, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
