+++ test/test_motor_collection.py
@@ -27,6 +27,7 @@ from tornado.concurrent import Future
 from tornado.testing import gen_test
 
 import motor
+import motor.core
 import test
 from test import MotorTest, assert_raises, version, SkipTest
 from test.utils import delay, skip_if_mongos
@@ -37,7 +38,7 @@ class MotorCollectionTest(MotorTest):
     def test_collection(self):
         # Test that we can create a collection directly, not just from
         # MotorClient's accessors
-        collection = motor.MotorCollection(self.db, 'test_collection')
+        collection = motor.core.MotorCollection(self.db, 'test_collection')
 
         # Make sure we got the right collection and it can do an operation
         self.assertEqual('test_collection', collection.name)
@@ -50,7 +51,10 @@ class MotorCollectionTest(MotorTest):
         # so this is prohibited.
         self.assertRaises(
             TypeError,
-            motor.MotorCollection, self.db, 'test_collection', capped=True)
+            motor.core.MotorCollection,
+            self.db,
+            'test_collection',
+            capped=True)
 
     @gen_test
     def test_dotted_collection_name(self):
@@ -378,7 +382,7 @@ class MotorCollectionTest(MotorTest):
         tmp_mr = yield collection.map_reduce(map_fn, reduce_fn, 'tmp_mr')
 
         self.assertTrue(
-            isinstance(tmp_mr, motor.MotorCollection),
+            isinstance(tmp_mr, motor.core.MotorCollection),
             'map_reduce should return MotorCollection, not %s' % tmp_mr)
 
         result = yield tmp_mr.find().sort([('_id', 1)]).to_list(length=1000)
@@ -472,7 +476,7 @@ class MotorCollectionTest(MotorTest):
 
         @gen.coroutine
         def f(cursor):
-            self.assertTrue(isinstance(cursor, motor.MotorCommandCursor))
+            self.assertTrue(isinstance(cursor, motor.core.MotorCommandCursor))
 
             while (yield cursor.fetch_next):
                 docs.append(cursor.next_object())
